feat:完成每日景点和人物

+++ lib_room/src/main/java/com/gp/room/entity/SceneryInfo.kt
@@ -13,7 +13,7 @@ import kotlinx.android.parcel.Parcelize
 @Entity(tableName = TABLE_SCENERY_LIST)
 data class SceneryInfo(
     @PrimaryKey(autoGenerate = false)
-    var id : Long = 0,
+    var id : Int = 0,
 
     @ColumnInfo(name = "nameVi", defaultValue = "")
     var nameVi : String?,
